Remove GM_addStyle
Since there's no need to bypass CSP restriction. Using GM_addStyle
and inserting a <style> element are almost the same for the result.

Greasemonkey v4+ seems not going to support GM_addStyle, so just
insert the CSS to the document without GM API and it works fine.

`@grant none` is assumed.